+++ modules/docker.nix
@@ -1,4 +1,10 @@
 {
+  darwinModule = { config, user, pkgs, ... }: {
+    environment.systemPackages = builtins.attrValues {
+      inherit (pkgs) colima docker-client docker-compose;
+    };
+  };
+
   nixosModule = { user, pkgs, ... }: {
     environment.systemPackages = builtins.attrValues {
       # Uses podman-compose instead of docker-compose